Study Buddies is an Android Application
The Study Buddies app will allow students to find a group with common
study interests and enable them to reserve a room in the NCSU Hunt
library.

The whole database is created and managed using Sqlite

When student start the app they required to login .On the main page
students will be able to create a new group as well as search for
existing groups to join using certain search parameters. The creator of
the group will be the admin and will reserve the right to add or delete
any member to the group. Once the study group is finalized, the students
will then be directed to NCSU Hunt Library's Reserve a room page where
they can reserve a room for the study group. Each Member of the group is
notified the timings of the scheduled meet via email.

Language : JAVA, Sqlite.  Platform: Android Studio.